added functionality for electricity and solar power
Added calculations for electricity and solar power.  A value for
sun_brightness was added to the System class.  Brightness and the
distance are the beginnings of the calculatinos for solar power.
Electricity is now simply something that is provided by buildings.  In
accordance with these changes the sun and electricity keys have been
removed from the Resources class since they are not calculated that way.