model building: Fleshed out more tests for the Mine and Solar Power Plant types
The basic assumption made by the _compare and value tests - i.e. that as
the level of a building goes up, the relative value of the building
increases. This may be true for the investment required to build the
building but some buildings, like the Mine, will have a negative effect
overall. The tests need to be updated to handle this. Until this
assumption is fixed, a number of tests will fail.