unittest: make lib model object attr test more permissive
The test expected all attributes to be present in the output of the
__str__ method in the form "attr: ", however some attributes look better
shown on the next line so are present in the form "attr:\n". This
implementation was chosen because checking for only "attr:" lets the
output look sloppy: "foo:3,bar:baz" vs "foo: 3, bar: baz"

Pedantic? Yes, get over it.