model planet: added name to constructor debug message
I had one test failure where 10 names were generated but the test only
counted 9 in the set. This indicates that namegen can produce duplicate
names. I need to decide if this is okay or if some preventative measures
need to be taken. Adding the name to the log message will let me double
check the error next time it happens.