tests: tweaked runtest to run faster and give better output
Removed "--failed" and "--with-id" flags because they aren't really
needed if the tests are run in parallel.  Also added the flag
"--detailed-errors" to help with investigating errors.

Finally, I added some stream filtering to the no-options execution of
the script. This way passed tests are not printed out but skipped and
failed tests are.