tests building: added tests for are_requirements_met method
The building and research requirements can't really be tested at this
time because there aren't any buildings with either type of requirement
defined at this point. However, I am considering adding a building
requirment of the same type but with one level less explicitly instead
of having that as an implicit requirement.