model user: allow user with no planets to be created
When loading the game state, we need to be able to create empty
instances of the objects and then apply the state from the save file.
This change isn't quite ideal still but it achieves the minimum goal.
More refactoring work is needed for the game state to work completely.