cmdline: moved game state business logic out of front end packages
There is an interesting divide/integration point between the Front End
and the internal game model. Knowledge of the internals of the game
model should be kept to the game model/engine code so that the Front End
doesn't need to handle the details.  At the same time the game
model/engine should not impose any User Interface requirements so that
the Front End can be concievably replaced with different interfaces in
the future.