model galaxy: refactored get/set state to call get/set state on children
In order to ensure the game state is fully converted into python
primitives, the game objects should be calling the get/set state methods
of any child game objects that are not themselves primitive already. A
test should be added to the Engine tests to check that all state
contents boil down to a nested structure of python primitives.