Use generator comprehension over list comprehensions where possible
List comprehensions must fully construct the list and save it in memory
before passing it into a method, such as in
"sum([x for x in range(10)])".  Using a generator expression like,
"sum(x for x in range(10))", no additional memory will be used until the
argument is referenced at which point the generator will return a value.
This saves a bit of memory and a few CPU cycles.