cmdline: removed unused commands 'planet' and 'user'
The existing planet and user commands may not fit with the command
interface design that is currently a work in progress.  Since this code
is not being used, and is hence getting stale, I'll just remove it until
I have a better idea of what each command should be doing.