model: add decorator that will trigger delayed actions
The idea is that model object instance methods can be decorated to
trigger the delayed action procedure before the method is executed.
Thus, when updating the resources on a planet, as an example, we can
take into account any delayed action - such as building construction -
that has yet to be completed but has passed its trigger time.