resources: define behaviour for subtraction cases resulting in negative values
The resources object is a representation of an amount of resources. Thus
negative values do not make any sense.  However, it is necessary to be
able to calculate differences between resources.  In the event that a
subtraction results in negative values an exception is thrown to
identify the fact that the action cannot proceed.  However, for code
that still needs to know what the result of that subtraction are, the
absolute value of the results are stored in the attribute "defecit" on
the NotSufficientResourcesError exception that gets raised.